Split up files and remove circular imports
